chore(clerk-js): Introduce ErrorCard and refactor HavingTrouble card
